using a single panzoom component to handle zoom for all image views
