+++ renderer/IndexImage/Blend.css
@@ -1,8 +1,3 @@
-.blend-zoom {
-  width: 100%;
-  height: 100%;
-}
-
 .blend {
   position: relative;
   width: var(--width, '100%');
