+++ renderer/IndexImage/Tolerance.css
@@ -1,8 +1,3 @@
-.tolerance-zoom {
-  width: 100%;
-  height: 100%;
-}
-
 .tolerance {
   position: relative;
   width: var(--width, '100%');
