+++ renderer/IndexImage/Tolerance.js
@@ -1,10 +1,10 @@
-const Panzoom = require('@panzoom/panzoom');
 const { html, css, useState, useEffect, useRef } = require('../tools/ui.js');
 const { tolerance, computeTolerance } = require('../tools/image-diff.js');
 
 css('./Tolerance.css');
 
 const Toolbar = require('../Toolbar/Toolbar.js');
+const Panzoom = require('./Panzoom.js');
 
 const KEY = 'tolerance';
 
@@ -13,10 +13,10 @@ const setVar = (elem, name, value) => elem.style.setProperty(`--${name}`, value)
 const toBackground = url => `url(${JSON.stringify(url)})`;
 
 function Tolerance({ left, right, buttons, cache }) {
-  const zoom = useRef(null);
   const view = useRef(null);
   const renderPromise = useRef(null);
   const [background, setBackground] = useState(toBackground(left));
+  const [zoomElem, setZoomElem] = useState(null);
 
   const [threshold, setThreshold] = useState(0.05);
 
@@ -32,31 +32,9 @@ function Tolerance({ left, right, buttons, cache }) {
     const ctx = view.current.getContext('2d');
     ctx.putImageData(new ImageData(data.output, data.width, data.height), 0, 0);
 
-    const box = view.current.getBoundingClientRect();
-    const win = zoom.current.getBoundingClientRect();
-
-    const startScale = Math.min(win.width / box.width, win.height / box.height, 1) * 0.98;
-    const startX = -((box.width / 2) - (win.width / 2));
-    const startY = -((box.height / 2) - (win.height / 2));
-
-    if (zoom.current.__x_panzoom) {
-      return zoom.current.__x_panzoom;
+    if (zoomElem !== view.current) {
+      setZoomElem(view.current);
     }
-
-    const panzoom = zoom.current.__x_panzoom = Panzoom(zoom.current, {
-      maxScale: 4,
-      startScale,
-      startX,
-      startY
-    });
-
-    zoom.current.addEventListener('wheel', panzoom.zoomWithWheel);
-
-    panzoom.__x_destroy = () => {
-      zoom.current.setAttribute('style', {});
-      zoom.current.removeEventListener('wheel', panzoom.zoomWithWheel);
-      panzoom.destroy();
-    };
   };
 
   useEffect(() => {
@@ -139,9 +117,9 @@ function Tolerance({ left, right, buttons, cache }) {
   return html`
     <${Toolbar}>${viewButtons}<//>
     <div class=main>
-      <div class="tolerance-zoom" ref=${zoom}>
+      <${Panzoom} view=${zoomElem}>
         <canvas class="tolerance" ref=${view}></canvas>
-      </div>
+      <//>
     </div>
   `;
 }
